added heartbeating
